refactor(on-demand-entries): extract method: setupPing
